utils.cleanup() was failing if cache directory was already removed

+++ lib/utils.py
@@ -38,6 +38,8 @@ DEFAULT_SINGULARITY_IMAGE = (
 
 
 def cleandir(d: str) -> None:
+    if not os.path.exists(d):
+        return
     with os.scandir(d) as it:
         for entry in it:
             os.unlink(f"{d}/{entry.name}")
@@ -46,7 +48,7 @@ def cleandir(d: str) -> None:
 
 def cleanup(varg: Dict[str, Any]) -> None:
     """cleanup submit directory etc."""
-    os.chdir(f'{varg["submitdir"]}/..')
+    os.chdir(os.path.dirname(f'{varg["submitdir"]}'))
     cleandir(varg["submitdir"])
     # now clean up old submit directories that weren't
     # cleaned up by jobsub_submit at the time
